+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 HTTP client extension Change Log
 2.0.15 under development
 ------------------------
 
-- no changes in this release.
+- Bug #224: Parse content when it is not an empty string (pawmaster)
 
 
 2.0.14 August 09, 2021